Extreme Difficulty Mode
An extra game mode (this is not asked for in the project document) is added. Since the algorithm behind the regular modes work randomly, the best path is not guaranteed. However, Extreme Difficulty Mode guarantees the best move for Computer. Note: Needs to be improved, sometimes throws 'ouf of index' errors, and squared 'unsquareable' areas a couple of times